requirements.txt
langchain
sentence_transformers
faiss-cpu
chromadb
python-multipart
uvicorn
BeautifulSoup4
InstructorEmbedding
httpx
python-dotenv